Полностью переделал обращение к другим сервисам и дописал весь TimetableService, но пока запросы с временем не работают

+++ TimetableService/Services/Timetables/TimetableServiceImpl.cs
@@ -1,5 +1,9 @@
 ﻿using AutoMapper;
+using AutoMapper.QueryableExtensions;
 using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
+using System.Net.Http.Headers;
+using TimetableService.Configuration;
 using TimetableService.Data;
 using TimetableService.Models.Appointments;
 using TimetableService.Models.Timetables;
@@ -11,15 +15,33 @@ public class TimetableServiceImpl : ITimetableService
 {
     private readonly DataContext _dataContext;
     private readonly IMapper _mapper;
+    private readonly HttpClient _httpClient;
+    private readonly IHttpContextAccessor _httpContextAccessor;
+    private readonly HttpClientRequestUri _httpClientRequestUri;
+    private readonly AppServices _services;
 
-    public TimetableServiceImpl(DataContext dataContext, IMapper mapper)
+    public TimetableServiceImpl(DataContext dataContext, IMapper mapper, IHttpClientFactory httpClientFactory, IHttpContextAccessor httpContextAccessor, HttpClientRequestUri httpClientRequestUri, AppServices services)
     {
         _dataContext = dataContext;
         _mapper = mapper;
+        _httpClient = httpClientFactory.CreateClient();
+        _httpContextAccessor = httpContextAccessor;
+        _httpClientRequestUri = httpClientRequestUri;
+        _services = services;
     }
 
     public async Task AddRecordToTimetable(AddTimetableRecordDTO dto)
     {
+        if (!(await DoesDoctorExists(dto.DoctorId)))
+        {
+            throw new ApplicationException("Нет такого доктора");
+        }
+
+        if (!(await DoesHospitalRoomExists(dto.HospitalId, dto.Room)))
+        {
+            throw new ApplicationException("Нет такого кабинета в больнице");
+        }
+
         var timetableRecord = new Timetable();
 
         _mapper.Map(dto, timetableRecord);
@@ -40,6 +62,11 @@ public class TimetableServiceImpl : ITimetableService
 
     public async Task DeleteDoctorRecordsFromTimetable(string doctorId)
     {
+        if (!(await DoesDoctorExists(doctorId)))
+        {
+            throw new ApplicationException("Нет такого доктора");
+        }
+
         await _dataContext.Timetables
             .Where(i => i.DoctorId == doctorId)
             .ExecuteDeleteAsync();
@@ -47,6 +74,11 @@ public class TimetableServiceImpl : ITimetableService
 
     public async Task DeleteHospitalRecordsFromTimetable(int hospitalId)
     {
+        if (!(await DoesHospitalExists(hospitalId)))
+        {
+            throw new ApplicationException("Нет такой больницы");
+        }
+
         await _dataContext.Timetables
             .Where(i => i.HospitalId == hospitalId)
             .ExecuteDeleteAsync();
@@ -59,23 +91,66 @@ public class TimetableServiceImpl : ITimetableService
             .ExecuteDeleteAsync();
     }
 
-    public Task<List<GetTimetableRecordDTO>> GetDoctorTimetable(string doctorId, string from, string to)
+    public async Task<List<GetTimetableRecordDTO>> GetDoctorTimetable(string doctorId, string from, string to)
     {
-        throw new NotImplementedException();
+        if (!(await DoesDoctorExists(doctorId)))
+        {
+            throw new ApplicationException("Нет такого доктора");
+        }
+
+        return await
+            _dataContext.Timetables
+                .Where(x => x.DoctorId == doctorId
+                    && x.From >= DateTime.Parse(from)
+                    && x.To <= DateTime.Parse(to))
+                .ProjectTo<GetTimetableRecordDTO>(_mapper.ConfigurationProvider)
+                .ToListAsync();
     }
 
-    public Task<List<GetTimetableRecordDTO>> GetHospitalRoomTimetable(int hospitalId, int roomId, string from, string to)
+    public async Task<List<GetTimetableRecordDTO>> GetHospitalRoomTimetable(int hospitalId, string room, string from, string to)
     {
-        throw new NotImplementedException();
+        if (!(await DoesHospitalRoomExists(hospitalId, room)))
+        {
+            throw new ApplicationException("Нет такого кабинета в больнице");
+        }
+
+        return await
+            _dataContext.Timetables
+                .Where(x => x.HospitalId == hospitalId && x.Room == room
+                    && x.From >= DateTime.Parse(from)
+                    && x.To <= DateTime.Parse(to))
+                .ProjectTo<GetTimetableRecordDTO>(_mapper.ConfigurationProvider)
+                .ToListAsync();
     }
 
-    public Task<List<GetTimetableRecordDTO>> GetHospitalTimetable(int hospitalId, string from, string to)
+    public async Task<List<GetTimetableRecordDTO>> GetHospitalTimetable(int hospitalId, string from, string to)
     {
-        throw new NotImplementedException();
+        if (!(await DoesHospitalExists(hospitalId)))
+        {
+            throw new ApplicationException("Нет такой больницы");
+        }
+
+        return await
+            _dataContext.Timetables
+                .Where(x => x.HospitalId == hospitalId
+                    && x.From >= DateTime.Parse(from)
+                    && x.To <= DateTime.Parse(to))
+                .ProjectTo<GetTimetableRecordDTO>(_mapper.ConfigurationProvider)
+                .ToListAsync();
     }
 
     public async Task UpdateRecordFromTimetable(int id, UpdateTimetableRecordDTO dto)
     {
+        if (!(await DoesDoctorExists(dto.DoctorId)))
+        {
+            throw new ApplicationException("Нет такого доктора");
+        }
+
+        if (!(await DoesHospitalRoomExists(dto.HospitalId, dto.Room)))
+        {
+            throw new ApplicationException("Нет такого кабинета в больнице");
+        }
+
         var timetableRecord = await
             _dataContext.Timetables
                 .Include(i => i.Appointments)
@@ -114,5 +189,64 @@ public class TimetableServiceImpl : ITimetableService
         await _dataContext.SaveChangesAsync();
 
         await tr.CommitAsync();
+    }
+
+    private async Task<bool> DoesDoctorExists(string doctorId)
+    {
+        var doctorRequest = new HttpRequestMessage()
+        {
+            RequestUri = new Uri(_services.AuthService + _httpClientRequestUri.GetDoctor.Replace("{id}", doctorId)),
+            Method = HttpMethod.Get,
+        };
+
+        doctorRequest.Headers.Authorization = new AuthenticationHeaderValue("Bearer", _httpContextAccessor.HttpContext?.Items["Token"]?.ToString());
+
+        var response = await _httpClient.SendAsync(doctorRequest);
+
+        return response.IsSuccessStatusCode;
+    }
+
+    private async Task<bool> DoesHospitalRoomExists(int hospitalId, string room)
+    {
+        var hospitalRequest = new HttpRequestMessage()
+        {
+            RequestUri = new Uri(_services.HospitalService + _httpClientRequestUri.GetHospitalRooms.Replace("{id}", hospitalId.ToString())),
+            Method = HttpMethod.Get,
+        };
+
+        hospitalRequest.Headers.Authorization = new AuthenticationHeaderValue("Bearer", _httpContextAccessor.HttpContext?.Items["Token"]?.ToString());
+
+        var response = await _httpClient.SendAsync(hospitalRequest);
+
+        if (response.IsSuccessStatusCode)
+        {
+            var rooms = JsonConvert.DeserializeObject<List<string>>(await response.Content.ReadAsStringAsync());
+
+            if (rooms == null || !rooms.Where(i => i == room).Any())
+            {
+                return false;
+            }
+
+            return true;
+        }
+        else
+        {
+            return false;
+        }
+    }
+
+    private async Task<bool> DoesHospitalExists(int hospitalId)
+    {
+        var hospitalRequest = new HttpRequestMessage()
+        {
+            RequestUri = new Uri(_services.HospitalService + _httpClientRequestUri.GetHospital.Replace("{id}", hospitalId.ToString())),
+            Method = HttpMethod.Get,
+        };
+
+        hospitalRequest.Headers.Authorization = new AuthenticationHeaderValue("Bearer", _httpContextAccessor.HttpContext?.Items["Token"]?.ToString());
+
+        var response = await _httpClient.SendAsync(hospitalRequest);
+
+        return response.IsSuccessStatusCode;
     }
 }
